[wpt] Fix the installation of ChromeDriver stable
This was accidentally regressed in #24702 due to lack of comments or
tests, both of which are also added in this change.

Fixes #24753.
